Clean build with gcc 3.3 on debian unstable
git-svn-id: https://svn.apache.org/repos/asf/xml/security/trunk@350996 13f79535-47bb-0310-9956-ffa450edef68

+++ c/src/tools/checksig/InteropResolver.cpp
@@ -125,7 +125,7 @@ X509 * InteropResolver::nextFile2Cert(void) {
 #if defined(_WIN32)
 		res = _findnext(m_handle, &m_finder);
 #else
-		if (m_fcount == m_globbuf.gl_pathc)
+		if (m_fcount == (int) m_globbuf.gl_pathc)
 			res = -1;
 		else
 			res = 0;
